+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [[11.1.4]](https://github.com/Clarifai/clarifai-python/releases/tag/11.1.4) - [PyPI](https://pypi.org/project/clarifai/11.1.4/) - 2025-02-12
+
+### Changed
+
+ - Introduce 3 times when you can download checkpoints [(#515)] (https://github.com/Clarifai/clarifai-python/pull/515)
+
+## [[11.1.3]](https://github.com/Clarifai/clarifai-python/releases/tag/11.1.3) - [PyPI](https://pypi.org/project/clarifai/11.1.3/) - 2025-02-11
+
+### Changed
+
+ - Fix dependency parsing [(#514)] (https://github.com/Clarifai/clarifai-python/pull/514)
+
+## [[11.1.2]](https://github.com/Clarifai/clarifai-python/releases/tag/11.1.2) - [PyPI](https://pypi.org/project/clarifai/11.1.2/) - 2025-02-10
+
+### Changed
+
+ - User new base images and fix clarifai version [(#513)] (https://github.com/Clarifai/clarifai-python/pull/513)
+
 ## [[11.1.1]](https://github.com/Clarifai/clarifai-python/releases/tag/11.1.1) - [PyPI](https://pypi.org/project/clarifai/11.1.1/) - 2025-02-06
 
 ### Changed